优化代码

+++ tplstyle/uninstall.php
@@ -4,6 +4,10 @@
  * @author WiFeng
  */
 
+if(!defined('IN_DISCUZ') || !defined('IN_ADMINCP')) {
+	exit('Access Denied');
+}
+
 $sql = "DROP TABLE IF EXISTS `pre_plugin_tplstyle`;";
 
 runquery($sql);
